Move chart component to container components directory
